refactor(client): create entryPoints dir
