+++ packages/webpack/src/webpack.ssr.config.js
@@ -5,7 +5,6 @@ const { join, resolve } = require('path');
 function ssr({ clientBasePath }) {
   return {
     mode: 'none',
-    // entry: [join(clientBasePath, '/src/ServerApp.js')],
     target: 'node',
     devtool: false,
     output: {
